% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ps.R
\name{get_ps}
\alias{get_ps}
\title{Get lorem ipsum}
\usage{
get_ps(spatial = TRUE)
}
\arguments{
\item{spatial}{a logical value. Return data of class \code{sf}? Defaults to \code{TRUE}}
}
\value{
data lorem ipsum
}
\description{
lorem ipsum
}
\author{
Jeppe Vierø
}
