% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_baselayer.R
\name{read_baselayer}
\alias{read_baselayer}
\title{Get Denmark baselayer shapefile}
\usage{
read_baselayer(as_sf = TRUE)
}
\arguments{
\item{as_sf}{A logical value. Return data as an object of class \code{sf}? Defauls to \code{TRUE}. If \code{FALSE}, data is imported as a GEOJSON of class \code{geomultipolygon}}
}
\value{
data lorem ipsum
}
\description{
lorem ipsum
}
\author{
Jeppe Vierø
}
