% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin.R
\name{get_admin}
\alias{get_admin}
\title{Get lorem ipsum}
\usage{
get_admin(spatial = TRUE, level = NULL)
}
\arguments{
\item{spatial}{a logical value. Return data of class \code{sf}? Defaults to \code{TRUE}}

\item{level}{lorem ipsum}
}
\value{
data lorem ipsum
}
\description{
lorem ipsum
}
\author{
Jeppe Vierø
}
