% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ps.R
\name{get_ps}
\alias{get_ps}
\title{Get lorem ipsum}
\usage{
get_ps(spatial = TRUE, level = NULL, crs_transform = FALSE, new_crs = NULL)
}
\arguments{
\item{spatial}{a logical value. Return data of class \code{sf}? Defaults to \code{TRUE}}

\item{level}{lorem ipsum}

\item{crs_transform}{Transform the CRS? Default is \code{FALSE}, which maintains the CRS at 4326.}

\item{new_crs}{A valid EPSG code. Required when \code{crs_transform = TRUE}. Transform the CRS by specifying a valid EPSG-code (see \code{rgdal::make_EPSG()$code} for a list of valid EPSG codes)}
}
\value{
data lorem ipsum
}
\description{
lorem ipsum
}
\author{
Jeppe Vierø
}
